Update Auth Structure
