+++ src/app/(site)/(authorized)/(app)/settings/actions.ts
@@ -9,6 +9,8 @@ import {
   PasswordChangeRequestSchema,
 } from "@/requests/password_change_request";
 import { UserRepository } from "@/repositories/user_repository";
+import { PasswordResetRepository } from "@/repositories/password_reset_repository";
+import { AuthService } from "@/services/auth_service";
 import { auth } from "@/libraries/auth";
 import { revalidatePath } from "next/cache";
 
@@ -47,7 +49,9 @@ export async function updateProfile(
     }
 
     const userRepository = new UserRepository();
-    await userRepository.updateProfile(
+    const passwordResetRepository = new PasswordResetRepository();
+    const authService = new AuthService(userRepository, passwordResetRepository);
+    await authService.updateProfile(
       session.user.id,
       validatedInput.data.name,
       validatedInput.data.email
@@ -87,7 +91,9 @@ export async function changePassword(
     }
 
     const userRepository = new UserRepository();
-    await userRepository.changePassword(
+    const passwordResetRepository = new PasswordResetRepository();
+    const authService = new AuthService(userRepository, passwordResetRepository);
+    await authService.changePassword(
       session.user.id,
       validatedInput.data.currentPassword,
       validatedInput.data.newPassword
