+++ src/app/(site)/(unauthorized)/auth/forgot-password/actions.ts
@@ -6,6 +6,8 @@ import {
 } from "@/requests/forgot_password_request";
 import { InvalidInput, Success } from "@/constants/auth";
 import { UserRepository } from "@/repositories/user_repository";
+import { PasswordResetRepository } from "@/repositories/password_reset_repository";
+import { AuthService } from "@/services/auth_service";
 
 export async function forgotPasswordAction(
   rawInput: ForgotPasswordRequest
@@ -16,8 +18,10 @@ export async function forgotPasswordAction(
       return InvalidInput;
     }
 
-    const authRepository = new UserRepository();
-    await authRepository.postForgotPassword(validatedInput.data);
+    const userRepository = new UserRepository();
+    const passwordResetRepository = new PasswordResetRepository();
+    const authService = new AuthService(userRepository, passwordResetRepository);
+    await authService.forgotPassword(validatedInput.data);
 
     return Success;
   } catch (error) {
