+++ src/repositories/password_reset_repository.ts
@@ -4,11 +4,6 @@ import {
 } from "@/models/password_reset";
 import { PrismaRepository } from "./prisma_repository";
 import { PasswordReset } from "@/models/password_reset";
-import {
-  generateResetToken,
-  createTokenExpiry,
-  isTokenExpired,
-} from "@/libraries/reset_token";
 
 export class PasswordResetRepository extends PrismaRepository<
   typeof PasswordResetSchema
@@ -17,60 +12,6 @@ export class PasswordResetRepository extends PrismaRepository<
     super(PasswordResetSchema, "passwordReset", transformPrismToModel, [
       "token",
     ]);
-  }
-
-  async createResetToken(userId: string): Promise<PasswordReset> {
-    const token = await generateResetToken();
-    const expiresAt = createTokenExpiry();
-    const now = new Date();
-
-    const resetData = {
-      userId,
-      token,
-      expiresAt,
-      usedAt: null,
-      createdAt: now,
-      updatedAt: now,
-    };
-
-    return this.create(resetData);
-  }
-
-  async findValidToken(token: string): Promise<PasswordReset | null> {
-    try {
-      const resetTokens = await this.get(
-        0,
-        1,
-        undefined,
-        undefined,
-        undefined,
-        [
-          { column: "token", operator: "=", value: token },
-          { column: "usedAt", operator: "=", value: null },
-        ]
-      );
-
-      if (resetTokens.data.length === 0) {
-        return null;
-      }
-
-      const resetToken = resetTokens.data[0];
-
-      if (isTokenExpired(resetToken.expiresAt)) {
-        return null;
-      }
-
-      return resetToken;
-    } catch (error) {
-      console.error("Error finding valid token:", error);
-      return null;
-    }
-  }
-
-  async markTokenAsUsed(tokenId: string): Promise<void> {
-    await this.update(tokenId, {
-      usedAt: new Date(),
-    });
   }
 
   async deleteUserTokens(userId: string): Promise<void> {
@@ -89,30 +30,6 @@ export class PasswordResetRepository extends PrismaRepository<
       }
     } catch (error) {
       console.error("Error deleting user tokens:", error);
-    }
-  }
-
-  async cleanupExpiredTokens(): Promise<void> {
-    try {
-      const now = new Date();
-      const expiredTokens = await this.get(
-        0,
-        1000,
-        undefined,
-        undefined,
-        undefined,
-        [{ column: "expiresAt", operator: "<", value: now }]
-      );
-
-      for (const token of expiredTokens.data) {
-        await this.delete(token.id);
-      }
-
-      console.log(
-        `Cleaned up ${expiredTokens.data.length} expired password reset tokens`
-      );
-    } catch (error) {
-      console.error("Error cleaning up expired tokens:", error);
     }
   }
